Separate plotting (#3117)
* Move plot.py to directory
