+++ demos/poisson/poisson_mixed.py.rst
@@ -147,7 +147,7 @@ visualisation tool such as `ParaView <http://www.paraview.org/>`__ ::
   File("poisson_mixed.pvd").write(u)
 
 We could use the built in plot function of firedrake by calling
-:func:`plot <firedrake.plot.tripcolor>` to plot a surface graph. Before that,
+:func:`plot <firedrake.pyplot.tripcolor>` to plot a surface graph. Before that,
 matplotlib.pyplot should be installed and imported::
 
   try:
@@ -156,6 +156,7 @@ matplotlib.pyplot should be installed and imported::
     warning("Matplotlib not imported")
 
   try:
+    from firedrake.pyplot import tripcolor
     fig, axes = plt.subplots()
     colors = tripcolor(u, axes=axes)
     fig.colorbar(colors)
