+++ firedrake/pyplot/__init__.py
@@ -1,0 +1,10 @@
+from .mpl import (
+    plot, triplot, tricontourf, tricontour, trisurf, tripcolor, quiver,
+    streamplot, FunctionPlotter
+)
+from .pgf import pgfplot
+
+__all__ = [
+    "plot", "triplot", "tricontourf", "tricontour", "trisurf", "tripcolor",
+    "quiver", "streamplot", "FunctionPlotter", "pgfplot"
+]
